dynamic-tee-vsink: Convert to playbin.
This attempts to use a playbin instead of a decodebin, and only attempts
to add a single new videosink for now (no removals). A new 'videosink' bin
is created, which contains the tee and all the rest. The videosink bin
is then installed as the playbin's videosink.

At the moment, this plays audio only; the new videosink never appears on
screen for some reason.